FINAL-1/almacen/requirements.txt
flask==3.0.0
flask_swagger_ui==4.11.1
PyYAML==6.0.1
